Development/messaging move (#196)
* [MESSAGING] Adapt the messaging to the moved to messaging code.

* [MESSAGECONTROL] Print/flus/send all message types (Logging/Tracing and WarningReporting)

* [MESSAGECONTROL] Oops a copy paste issue :-)

* [MEESAGING] New method usage.

+++ OpenCDMi/FrameworkRPC.cpp
@@ -329,7 +329,7 @@ namespace Plugin {
                                 sampleInfo.ivLength = IVKeyLength();
                                 sampleInfo.keyId = const_cast<uint8_t *>(KeyId(sampleInfo.keyIdLength));
                                 sampleInfo.subSample = const_cast<CDMi::SubSampleInfo *>(SubSamples());
-                                sampleInfo.subSampleCount = SubSampleLength();
+                                sampleInfo.subSampleCount = static_cast<uint8_t>(SubSampleLength());
 
                                 uint16_t width = 0, height = 0;
                                 uint8_t type = 0;
